add example on how to use key binding context
key binding context are useful when you want to use the same shortcut but
do differents things based on the context. In the example, you can see
that the shortcut `ctrl+alt+d` will open "devdocs.io" if you are editing a
Python file, and the same shortcut will open "developer.mozilla.org" if
you are editing a JavaScript file. This way, you don't have to bind and to
remember X shortcuts for the X languages you know :)